[bresenham] working

+++ 2DGridFOV/2DGridFOV/Models/Equation.cs
@@ -44,11 +44,11 @@ namespace _2DGridFOV.Models
             this._b = a[1] - this._m*a[0]; //b = y - mx
         }
 
-        public int GetXValue(int y) // y = mx + b => x = (y - b)/m
+        public int GetXValue(float y) // y = mx + b => x = (y - b)/m
         {
             return (int)((y - this._b) / this._m);
         }
-        public int GetYValue(int x)
+        public int GetYValue(float x)
         {
             return (int)(this._m * x + this._b);
         }
